implement tokenizer to simplify parser
