+++ Cargo.toml
@@ -17,7 +17,6 @@ crate-type = ['cdylib', 'rlib']
 [dependencies]
 serde = { version = "1", features = ["derive"] }
 serde_json = {version = "1", features = ["preserve_order"]}
-thiserror = "1"
 
 [dev-dependencies]
 criterion = "0.3"
