+++ src/dsl/error.rs
@@ -1,0 +1,11 @@
+#[derive(Debug)]
+pub struct ParseError {
+    pub(crate) pos: usize,
+    pub(crate) cause: Box<ParseErrorCause>,
+}
+
+#[derive(Debug)]
+pub enum ParseErrorCause {
+    UnexpectedEndOfInput,
+    UnexpectedChar(char),
+}
