+++ src/dsl/mod.rs
@@ -1,0 +1,13 @@
+// TODO: remove this when this module starts being used
+#![allow(dead_code)]
+
+mod ast;
+mod error;
+mod parser;
+mod token;
+mod tokenizer;
+#[cfg(test)]
+mod test;
+
+pub use error::ParseError;
+pub use ast::{Rhs, Lhs};
